Refactoring Auth for travels
